+++ project/src/types/auhtorise.ts
@@ -1,0 +1,13 @@
+export type AuhtoriseUser = {
+  avatarUrl: string;
+  email: string;
+  id: number;
+  isPro: boolean;
+  name: string;
+  token: string;
+}
+
+export type AuthData = {
+  login: string;
+  password: string;
+}
