+++ project/src/types/offer.ts
@@ -43,11 +43,3 @@ export type Comment = [{
   user: Host;
 }]
 
-export type AuhtoriseUser = {
-  avatarUrl: string;
-  email: string;
-  id: number;
-  isPro: boolean;
-  name: string;
-  token: string;
-}
